// apps/frontend/src/gen/vibesync/media/v1/media_pb.ts
export enum MediaKind {
  /**
   * @generated from enum value: MEDIA_KIND_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: MEDIA_KIND_AUDIO = 1;
   */
  AUDIO = 1,

  /**
   * @generated from enum value: MEDIA_KIND_VIDEO = 2;
   */
  VIDEO = 2,
}
// Retrieve enum metadata with: proto3.getEnumType(MediaKind)
proto3.util.setEnumType(MediaKind, "vibesync.media.v1.MediaKind", [
  { no: 0, name: "MEDIA_KIND_UNSPECIFIED" },
  { no: 1, name: "MEDIA_KIND_AUDIO" },
  { no: 2, name: "MEDIA_KIND_VIDEO" },
]);

/**
 * @generated from enum vibesync.media.v1.MediaSource
 */
export enum MediaSource {
  /**
   * @generated from enum value: MEDIA_SOURCE_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: MEDIA_SOURCE_PROVIDER = 1;
   */
  PROVIDER = 1,

  /**
   * @generated from enum value: MEDIA_SOURCE_UPLOAD = 2;
   */
  UPLOAD = 2,

  /**
   * @generated from enum value: MEDIA_SOURCE_CACHE = 3;
   */
  CACHE = 3,
}